+++ CHANGELOG.md
@@ -36,6 +36,11 @@
 
 ### Bug Fixes
 
+* Fix `unused_declaration` false positives for UIKit/WebKit delegate types and
+  optional protocol witness methods in subclasses.  
+  [leno23](https://github.com/leno23)
+  [#4706](https://github.com/realm/SwiftLint/issues/4706)
+
 * Avoid false positives in `prefer_self_in_static_references` for generic
   constraints and generic parameter bounds such as `where A: P` and `<A: P>`
   in classes and extensions.